Feb 6, 2025, 4:20 AM

+++ CT07_03/lesson3.py
@@ -12,5 +12,5 @@
 savings = 0
 while savings < 100:
     savings_today = input("How much you save today?")
-    savings = savings + int(savings_today)
+    savings = savings + float(savings_today)
 print("U suck. Noone likes u. Your free trial has ended, now you must pay " + str(savings) + " dollars.")
